+++ src/NepTrain/core/validation/__init__.py
@@ -1,0 +1,1 @@
+"""Non-fatal magnetic training data validators."""